# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "GoIM.Push.Service 服务器接口文档",
        "title": "GoIM.Push.Service Swagger",
        "termsOfService": "http://go-goim.github.io/",
        "contact": {
            "name": "Yusank",
            "url": "https://yusank.space",
            "email": "yusankurban@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/go-goim/core/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/push/v1",
    "paths": {
        "/push/v1/conn/ws": {
            "get": {
                "description": "websocket 长连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Conn"
                ],
                "summary": "websocket",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "invalid jwt cookie",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Meta": {
            "type": "object",
            "properties": {
                "extra": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "reason": {
                    "type": "string"
                }
            }
        }
    }
}